Refactor megadetector_json_to_csv.py
- encapsulate some functions/code that are shared between a few of these
  conversions scripts to refactor this script and prepare to refactor
  the others

- make iteration through image detection more pythonic

- moved a step that was being executed every time through a loop but
  didn't need to be out of the loop.
