+++ conversions/megadetector_json_to_csv.py
@@ -18,13 +18,16 @@ import csv
 import argparse
 import random
 
+from dataprep import bbox_to_pascal
+
 
 def main():
-    """Converts JSON data and writes to CSV
+    """Prepares a CSV file for camml TFLite training from MD JSON file
 
     Converts the JSON xywh bbox data to xyxy format and writes the
     data to a CSV in the proper format [set, class, file, xmin, ymin,
     '', '', xmax, ymax, '', ''].
+
     """
     # pylint: disable=locally-disabled, too-many-locals
     # Get command line arguments when running program
@@ -56,29 +59,19 @@ def main():
             # Set the file path
             image_path = img['file']
 
+            # Get the class label from the image file name
+            new_label = img['file'].strip('/').split('/')[-2]
+            
             if 'failure' in img.keys():
                 print(img['file'] + ' failed to access.\n')
             else:
-                for i in range(0, len(img['detections'])):
-                    # Convert xywh bbox to xmin, ymin, xmax, ymax bbox
-                    x_min, y_min, x_max, y_max = coco_to_pascal_voc(
-                        img['detections'][i]['bbox'][0],
-                        img['detections'][i]['bbox'][1],
-                        img['detections'][i]['bbox'][2],
-                        img['detections'][i]['bbox'][3])
-
-                    # Round to 4 digits to match CSV format
-                    x_min, y_min = round(x_min, 4), round(y_min, 4)
-                    x_max, y_max = round(x_max, 4), round(y_max, 4)
-
-                    # Set bbox coordinates to new values
-                    img['detections'][i]['bbox'][0] = x_min
-                    img['detections'][i]['bbox'][1] = y_min
-                    img['detections'][i]['bbox'][2] = x_max
-                    img['detections'][i]['bbox'][3] = y_max
-
+                for detection in img['detections']:
+                    # convert bbox to Pascal VOC
+                    bbox_to_pascal(detection['bbox'])
+                    
                     # Randomly set 80% of images to train,
                     # 10% to validation, and 10% to test. Currently all set
+                    # random.seed(42)  # for testing
                     rand_num = random.randint(1, 100)
                     set_type = ''
 
@@ -89,22 +82,19 @@ def main():
                     elif rand_num <= 100:
                         set_type = 'TEST'
 
-                    # Get the class name from the image file name
-                    category = img['file'].strip('/').split('/')[-2]
-
                     # Megadetector uses 3 categories 1-animal, 2-person,
                     # 3-vehicle, only the animal detections are needed
                     # Filters detections so only >= conf detections appear
-                    if (img['detections'][i]['category'] == '1'
-                            and img['detections'][i]['conf'] >= args.conf):
+                    if (detection['category'] == '1'
+                            and detection['conf'] >= args.conf):
                         csv_writer.writerow([set_type,
                                              image_path,
-                                             category,
-                                             img['detections'][i]['bbox'][0],
-                                             img['detections'][i]['bbox'][1],
+                                             new_label,
+                                             detection['bbox'][0],
+                                             detection['bbox'][1],
                                              None, None,
-                                             img['detections'][i]['bbox'][2],
-                                             img['detections'][i]['bbox'][3],
+                                             detection['bbox'][2],
+                                             detection['bbox'][3],
                                              None, None])
 
                 # To make images with no detections appear in the csv file
@@ -115,24 +105,6 @@ def main():
                                              None, None, None,
                                              None, None, None,
                                              None, None, None])
-
-
-def coco_to_pascal_voc(x_tl, y_tl, width, height):
-    """Convert Coco bounding box to Pascal Voc bounding box
-
-    Parameters:
-        x_tl (float):Float representing the min x coordinate of the bbox
-        y_tl (float):Float representing the min y coordinate of the bbox
-        width (float):Float representing the width of the bbox
-        height (float):Float representing the height of the bbox
-
-    Returns:
-        x_min (float):Float representing the min x coordinate of the bbox
-        y_min (float):Float representing the min y coordinate of the bbox
-        x_max (float):Float representing the max x coordinate of the bbox
-        y_max (float):Float representing the max y coordinate of the bbox
-    """
-    return [x_tl, y_tl, x_tl + width, y_tl + height]
 
 
 if __name__ == "__main__":
